Initial commit. All required features are complete.
```
COMP3606 PRACTICAL 3 (Intents)

1.  Create the Main screen for a Furniture ordering app.

2.  Create a second screen for the app. The second screen must receive some       data
from the main (first) screen. At the second screen, print the data received
from the first screen.

 3. Create a third screen. The third screen is called by the second screen to perform a
calculation based on the data from the second screen. There should be a button
to click to return to the second screen where the calculation result is
displayed.

 4. Add a new feature of your own.

 5. Ensure that your app has a few pictures.

End of Practical
```

For #4, I added background color changing to a random color and persisting that background color across all activities.